+++ src/tests/init_tests.py
@@ -20,10 +20,13 @@
 # IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 # CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+import sys
+import os.path
+sys.path.insert(0, os.path.abspath('..'))
 import hcpsdk
 
 # primary HCP
-P_HCP       = 'hcp1.snomis.local'
+P_HCP       = 'hcp.snomis.local'
 P_ADMIN     = 'admin.' + P_HCP
 P_TENANT    = 'm.' + P_HCP
 P_NS_GOOD   = 'n1.' + P_TENANT